Move DTOs to separate file
