+++ backend/Veloci.Logic/Services/Patreon/IPatreonService.cs
@@ -1,4 +1,5 @@
 using Veloci.Data.Domain;
+using Veloci.Logic.Services.Patreon.Models;
 
 namespace Veloci.Logic.Services;
