+++ backend/Veloci.Logic/Services/Patreon/PatreonApiClient.cs
@@ -2,6 +2,7 @@ using System.Net;
 using System.Net.Http.Headers;
 using System.Text.Json;
 using Microsoft.Extensions.Logging;
+using Veloci.Logic.Services.Patreon.Models;
 
 namespace Veloci.Logic.Services;
